pyproject.toml
[tool.black]
line-length = 88
target_version = ['py37', 'py38', 'py39', 'py310', "py311"]
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
    | \.tox
  )/
)
'''

[tool.isort]
line_length = 88
profile = "black"

[tool.poetry]
name = "openshift-ci-trigger"
version = "1.0.0"
description = "Trigger openshift CI job based on merged MR in Gitlab"
authors = ["Meni Yakove <myakove@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "*"
python-gitlab = "*"
PyYaml = "*"
PyGithub = "*"
gitpython = "^3.1.31"

[tool.poetry.group.dev.dependencies]
ipython = "*"
ipdb = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120

## Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".eggs",
    ".git",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "build",
    "dist",
    "node_modules",
    "venv",
]
